옵션 선택시 설정을 저장

+++ Journal/Sources/Controllers/EntryTableViewCell.swift
@@ -13,7 +13,7 @@ struct EntryTableViewCellModel {
     let environment: Environment
     
     var entryText: String { return entry.text }
-    var entryTextFont: UIFont { return UIFont.systemFont(ofSize: CGFloat(environment.settings.fontSize.rawValue)) }
+    var entryTextFont: UIFont { return UIFont.systemFont(ofSize: environment.settings.fontSize.rawValue) }
     var createdDateText: String { return DateFormatter.timeFormatter.string(from: entry.createdAt) }
     var ampmText: String { return DateFormatter.ampmFormatter.string(from: entry.createdAt) }
 }
