+++ Journal/Sources/Controllers/EntryViewControllerModel.swift
@@ -27,7 +27,7 @@ class EntryViewControllerModel {
     
     var textViewText: String? { return entry?.text }
     var textViewFont: UIFont {
-        return UIFont.systemFont(ofSize: CGFloat(environment.settings.fontSize.rawValue))
+        return UIFont.systemFont(ofSize: environment.settings.fontSize.rawValue)
     }
     
     private(set) var isEditing: Bool = false
